Fix specs

+++ app/controllers/invitations_controller.rb
@@ -1,19 +1,21 @@
 class InvitationsController < ApplicationController
   skip_filter :authenticate_user!, only: :accept
-  before_action :authorize_invitation, except: :accept
   before_action :find_active_invitation, only: :accept
   before_action :find_invitation, only: [:destroy, :resend]
 
   def new
+    authorize :invitation, :new?
     @invitation = Invitation.new
   end
 
   def index
+    authorize :invitation, :index?
     @invitations = current_organization.invitations
   end
 
   def create
     @invitation = current_member.invitations.build(invitation_params)
+    authorize @invitation
 
     if @invitation.save
       redirect_to new_invitation_path, notice: 'An invitation was created successfully'
@@ -23,6 +25,7 @@ class InvitationsController < ApplicationController
   end
 
   def resend
+    authorize @invitation
     if @invitation.send_invitation
       redirect_to invitations_path, notice: 'An invitation has been resent'
     else
@@ -31,6 +34,7 @@ class InvitationsController < ApplicationController
   end
 
   def destroy
+    authorize @invitation
     if @invitation.destroy
       redirect_to invitations_path, notice: 'An inventation was destroyed successfully'
     else
@@ -53,25 +57,20 @@ class InvitationsController < ApplicationController
   end
 
   private
+    def find_active_invitation
+      @invitation = Invitation.active.find_by(token: params[:token])
+      redirect_to root_path, alert: 'Bad invitation token' unless @invitation
+    end
 
-  def authorize_invitation
-    authorize :invitation, :owner_or_admin_for_record?
-  end
+    def find_invitation
+      @invitation = Invitation.find(params[:id])
+    end
 
-  def find_active_invitation
-    @invitation = Invitation.active.find_by(token: params[:token])
-    redirect_to root_path, alert: 'Bad invitation token' unless @invitation
-  end
+    def invitation_params
+      params.require(:invitation).permit(:email, :role)
+    end
 
-  def find_invitation
-    @invitation = Invitation.find(params[:id])
-  end
-
-  def invitation_params
-    params.require(:invitation).permit(:email, :role)
-  end
-
-  def user_params
-    params.require(:user).permit(:full_name, :password)
-  end
+    def user_params
+      params.require(:user).permit(:full_name, :password)
+    end
 end
